23-01-22 09:08

+++ src/main/java/dii2dam/movieApp/dao/ActorDaoInt.java
@@ -5,11 +5,11 @@ import java.util.List;
 import dii2dam.movieApp.models.Actor;
 
 public interface ActorDaoInt extends CommonDaoInt<Actor>{
-	  public List<Actor> searchByActorId(final int id);
-	  
-	  public List<Actor> searchByName(final String name);
-	  
-	  public List<Actor> searchBySurname(final String surname);
-	  
-	  public List<Actor> searchByBithdate(final String birthdate);
+  public List<Actor> searchByActorId(final Long id);
+  
+  public List<Actor> searchByName(final String name);
+  
+  public List<Actor> searchBySurname(final String surname);
+  
+  public List<Actor> searchByBithdate(final String birthdate);
 }
